refactor(menu): split menu and menuitem
